create DigitalOutputsClass as separate module
